requirements.txt
pytesseract
PyPDF2
Pillow
python-docx
python-pptx
openpyxl
pandas
anthropic
python-dotenv